ai-model/requirements.txt
tensorflow>=2.10.0
opencv-python>=4.8.0
numpy>=1.21.0
pillow>=9.0.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
flask>=2.2.0
requests>=2.28.0
python-dotenv>=0.19.0
